Solve all warnings in simulator and add button down events

+++ src/font.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 typedef struct {
-    int width;
+    uint8_t width;
     const uint8_t bitmap[5];
 } font_character_t;
